\name{gbm_wrapper}
\alias{gbm_wrapper}
\title{Wrapper for gbm selection and estimation}
\usage{
gbm_wrapper( target, data, distribution )
}
\description{
Wraps calls to gbm package for feature selection and model estimation.
}
\examples{
## NOT RUN:
## gbm_wrapper( target='COVID_NEW_ADM_CNT', data=covid )
}
