% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_wrapper.R
\name{gbm_wrapper}
\alias{gbm_wrapper}
\title{Function to fit a gbm model}
\usage{
gbm_wrapper(
  target,
  data,
  n.trees = 5000,
  shrinkage = 0.0025,
  interaction.depth = 5,
  distribution = "gaussian",
  ...
)
}
\arguments{
\item{target}{name of the column that is being predicted}

\item{data}{`data.frame` of predictors and the response variable}

\item{n.trees}{how many trees in the random forest}

\item{shrinkage}{learning rate}

\item{interaction.depth}{how many branches per tree}

\item{distribution}{statistical distribution of the response variable}
}
\value{
a fitted GBM model
}
\description{
This wraps up my most common options when building a gradient boosted random forest model with GBM
}
